Marker popup slot & color prop
This modifies the Marker component so that the props are always consistently passed when instantiating the `mapbox.Marker` class.

It also adds a `color` prop, which allows you to customize the default marker’s color.

Another slot has been added to the component.  The `popup` slot provides a quick way to create a popup that displays when the marker is clicked.  It allows passing markup, and the Vue events & template syntax all work wonderfully.

To wrap it up, I’ve added some documentation on Markers in `markers.md`.